Handle loading state globally with state in ui slice and App-Overlay
