+++ src/app/layout/index.ts
@@ -1,4 +1,5 @@
 import ToolBar from './toolbar/ToolBar'
 import Navigation from './navigation/Navigation'
+import Content from './content/Content'
 
-export { Navigation, ToolBar }
+export { Navigation, ToolBar, Content }
