+++ src/features/registration/store/selectors.ts
@@ -1,3 +1,2 @@
-export const isLoading = state => state.registration.isLoading
 export const getError = state => state.registration.error
 export const isSuccess = state => state.registration.isSuccess
