+++ src/features/registration/store/types.ts
@@ -1,5 +1,4 @@
 export interface RegistrationState {
-  isLoading: boolean
   isSuccess: boolean
   error?: object
 }
