+++ src/shared/components/LoadingOverlay/LoadingOverlay.styl
@@ -1,0 +1,15 @@
+.loading-screen
+  position relative
+
+  &__overlay
+    position absolute
+    height 100%
+    width 100%
+    background-color rgba(250, 250, 250, .8)
+    z-index 1
+
+  &__spinner
+    position absolute
+    top 50%
+    left 50%
+    transform translate(-50%, -50%)
