+++ src/shared/components/index.ts
@@ -1,1 +1,2 @@
 export * from './Form/form'
+export * from './LoadingOverlay/LoadingOverlay'
