+++ src/shared/hocs/index.ts
@@ -1,0 +1,1 @@
+export { withLoadingOverlay } from './withLoadingOverlay'
